requirements.txt
numpy>=1.26
scipy>=1.12
pandas>=2.2
pyarrow>=15.0
psycopg2-binary>=2.9
python-dotenv>=1.0
